Introducing BTDMatrix, a block-tridiagonal matrix.  It implements the Thomas algorithm to solve block-tridiagonal systems of equations.
NOTE:  Yes, I know that this is not the time to add new features.
However, tridiagonal matrices and the Thomas algorithm are important
components in line smoothers for anisotropic problems.  We will
need that soon for the AdaptHydroMod project, and therefore I would
like to have this matrix in the 2.0 release.

Also, this patch is not invasive at all: it just adds another matrix
class.

Please merge this to the release branch.



git-svn-id: https://svn.dune-project.org/svn/dune-istl/trunk@1158 8c7282f0-4411-0410-8e4f-cdd3e049afb8